Fix ошибок + доработка ASGI 	modified:   db.sqlite3 	modified:   forum/__pycache__/consumers.cpython-311.pyc 	modified:   forum/__pycache__/views.cpython-311.pyc 	modified:   forum/consumers.py # Доделано удаление + Добавлена развилка на отправку сообщения 	modified:   forum/templates/card/question_detail.html # onDelete usage 	modified:   forum/views.py

+++ forum/consumers.py
@@ -1,7 +1,7 @@
 from channels.generic.websocket import AsyncWebsocketConsumer
 import json
 from channels.db import database_sync_to_async
-from .models import Question
+from .models import Message, User, Question
 
 class QuestionConsumer(AsyncWebsocketConsumer):
     async def connect(self):
@@ -23,17 +23,38 @@ class QuestionConsumer(AsyncWebsocketConsumer):
     async def receive(self, text_data):
         text_data_json = json.loads(text_data)
         message_id = text_data_json['id']
-
-        # Удаление сообщения из базы данных
-        await self.delete_message(message_id)
-
-        # Отправка сообщения об удалении всем клиентам
-        await self.channel_layer.group_send(
-            self.room_group_name, 
-            {
-                'type': 'question_message',
-                'id': message_id,
-            }
+        
+        if "id" in text_data_json:
+            # Удаление сообщения из базы данных
+            await self.delete_message(message_id)
+            # Отправка сообщения об удалении всем клиентам
+            await self.channel_layer.group_send(
+                self.room_group_name, 
+                {
+                    'type': 'question_message',
+                    'id': message_id,
+                }
+            )
+        
+        if 'text' in text_data_json:
+            message_text = text_data_json['text']
+            message = self.create_message(message_text)
+            await self.channel_layer.group_send(
+                self.room_group_name,
+                {
+                    'type': 'send_message',
+                    'message': message_text,
+                    'id': message.id,
+                    'author': message.author.username
+                    }
+                )
+    
+    async def send_message(self, event):
+        await self.send(text_data=json.dumps({
+            "id": event['id'],
+            "message": event['message'],
+            "author": event['author']
+            })
         )
 
     async def question_message(self, event):
@@ -48,6 +69,18 @@ class QuestionConsumer(AsyncWebsocketConsumer):
     def delete_message(self, id):
         # Здесь логика удаления сообщения из базы данных
         try: 
-            Question.objects.get(pk=id).delete()
-        except Question.DoesNotExist:
-            pass
+            message = Message.objects.get(pk=id)
+            message.delete()
+        except Message.DoesNotExist:
+            print("Сообщения не существует")
+    
+    def create_message(self, text):
+        user = User.objects.get(username=self.scope["user"].username)
+        question = Question.objects.get(pk=self.question_id)
+
+        message = Message.objects.create(
+            author=user,
+            question=question,
+            text=text
+            )
+        return message
